+++ src/main/java/roomescape/presentation/MemberViewController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.presentation;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
