+++ src/main/java/roomescape/presentation/response/ReservationTimeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.controller.response;
+package roomescape.presentation.response;
 
 import java.time.LocalTime;
 import roomescape.service.result.ReservationTimeResult;
