+++ src/main/java/roomescape/presentation/response/ThemeResponse.java
@@ -1,9 +1,10 @@
-package roomescape.controller.response;
+package roomescape.presentation.response;
 
 import roomescape.service.result.ThemeResult;
 
 public record ThemeResponse(Long id, String name, String description, String thumbnail) {
     public static ThemeResponse from(final ThemeResult themeResult) {
-        return new ThemeResponse(themeResult.id(), themeResult.name(), themeResult.description(), themeResult.thumbnail());
+        return new ThemeResponse(themeResult.id(), themeResult.name(), themeResult.description(),
+                themeResult.thumbnail());
     }
 }
